Freecivx-server-java

+++ freecivx-server/src/main/java/net/freecivx/game/City.java
@@ -22,7 +22,7 @@ package net.freecivx.game;
 public class City {
 
     private String name;
-    private int owner;
+    private long owner;
     private long tile;
     private int size;
     private int style;
@@ -36,7 +36,7 @@ public class City {
     private int productionValue;
 
     // Constructor
-    public City(String name, int owner, long tile, int size, int style, boolean capital, boolean occupied, int walls,
+    public City(String name, long owner, long tile, int size, int style, boolean capital, boolean occupied, int walls,
                 boolean happy, boolean unhappy, String improvements, int productionKind, int productionValue) {
         this.name = name;
         this.owner = owner;
@@ -62,11 +62,11 @@ public class City {
         this.name = name;
     }
 
-    public int getOwner() {
+    public long getOwner() {
         return owner;
     }
 
-    public void setOwner(int owner) {
+    public void setOwner(long owner) {
         this.owner = owner;
     }
